Update README.md
PART 1:  Generate public/private key pairs with GPG.

Watch the video and follow along. When finished, take screenshot_1 as indicated in the video.



PART 2:  Encrypt a file using your public key and then decrypt the same file with your private key.

Watch the video and follow along. When finished, take screenshot_2 as indicated in the video.



PART 3:  Export your public key to a file and display it in its full length.

Watch the video and follow along. When finished, take screenshot_3 as indicated in the video.



PART 4:  Create a new user and have that user import the public key from PART 3.

Watch the video and follow along. When finished, take screenshot_4 as indicated in the video.



PART 5:  Decrypt a message with your private key that was encrypted by another user with your shared public key.

Watch the video and follow along. When finished, take screenshot_5 as indicated in the video.